fix: Game: moved saveEndGameStats to statistics TU
* Moved saveToFileEndGameStatistics() to anonymous namespace in
statistics TU.
