+++ src/headers/statistics.hpp
@@ -6,6 +6,10 @@
 #include <string>
 #include <tuple>
 
+namespace Scoreboard {
+struct Score;
+}
+
 namespace Statistics {
 struct total_game_stats_t {
   ull bestScore{};
@@ -18,8 +22,8 @@ struct total_game_stats_t {
 using load_stats_status_t = std::tuple<bool, total_game_stats_t>;
 
 load_stats_status_t loadFromFileStatistics(std::string filename);
-bool saveToFileEndGameStatistics(std::string filename, total_game_stats_t s);
 ull load_game_best_score();
+void saveEndGameStats(Scoreboard::Score finalscore);
 } // namespace Statistics
 
 std::istream &operator>>(std::istream &is, Statistics::total_game_stats_t &s);
